requirements.txt
# Core Deep Learning Framework
torch>=1.8.0
torchvision>=0.9.0

# Tensor Operations
einops>=0.4.0

# Image Processing
Pillow>=8.0.0
opencv-python>=4.5.0
scikit-image>=0.18.0

# Numerical Computing
numpy>=1.19.0

# Visualization
matplotlib>=3.3.0

# Model Summary
torchsummary>=1.5.1

# Progress Bar
tqdm>=4.60.0
